refactor: fix simplify review findings
- Extract duplicate formatRelativeTime to shared string-utils.ts
- Delete dead resolveSelectedWorkspaceId/resolveSelectedSessionId from app-store-utils
- Eliminate double buildSlashCommandSections call in use-slash-menu (use already-computed slashSuggestions)
- Remove redundant mapToRecord call in handleSessionEvent (markSessionViewed already handles it)
