+++ apps/desktop/src/App.tsx
@@ -8,6 +8,7 @@ import {
   type WorktreeRecord,
   type WorkspaceRecord,
 } from "./desktop-state";
+import { formatRelativeTime } from "./string-utils";
 import { ComposerPanel } from "./composer-panel";
 import type { ComposerSlashCommand } from "./composer-commands";
 import { desktopCommands, getDesktopCommandFromShortcut, type PiDesktopCommand } from "./ipc";
@@ -51,27 +52,6 @@ function useDesktopAppState() {
   }, []);
 
   return [snapshot, setSnapshot] as const;
-}
-
-function formatRelativeTime(value: string): string {
-  if (!value) {
-    return "";
-  }
-
-  const timestamp = Date.parse(value);
-  if (Number.isNaN(timestamp)) {
-    return value;
-  }
-
-  const diffMs = Date.now() - timestamp;
-  const diffMinutes = Math.max(0, Math.floor(diffMs / 60000));
-  if (diffMinutes < 1) return "now";
-  if (diffMinutes < 60) return `${diffMinutes}m`;
-  const diffHours = Math.floor(diffMinutes / 60);
-  if (diffHours < 24) return `${diffHours}h`;
-  const diffDays = Math.floor(diffHours / 24);
-  if (diffDays < 7) return `${diffDays}d`;
-  return new Date(timestamp).toLocaleDateString();
 }
 
 function updateSnapshot(
